templates/project_template/requirements.txt
pydantic-ai
openai
pgvector
psycopg2-binary
sqlalchemy
beautifulsoup4
requests
streamlit
python-dotenv
lxml